requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0

# LLM APIs
openai>=1.10.0
google-generativeai>=0.3.0
mistralai>=0.0.11

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Experiment Tracking
wandb>=0.16.0
tensorboard>=2.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# NLP Utilities
nltk>=3.8
spacy>=3.7.0
sentencepiece>=0.1.99

# Model Quantization (optional, for QLoRA)
bitsandbytes>=0.42.0

# French NLP
# python -m spacy download fr_core_news_md
